LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointercancel_touch-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
pointercancel test

Warning: this test works properly only for devices that have touchscreen

Test Description: This test checks if pointercancel event triggers.
Start touch over the black rectangle and then move your finger to scroll the page.


Pointer Events pointercancel Tests

The following pointer types were detected: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN pointercancel event received
